fix: set appropriate mime type for swagger UI.
