+++ external/plugins/swaggerui/impl/plugin.go
@@ -1,0 +1,88 @@
+package main
+
+import (
+	"fmt"
+	"github.com/hashicorp/go-hclog"
+	goplugin "github.com/hashicorp/go-plugin"
+	"github.com/imposter-project/imposter-go/external/handler"
+	"github.com/imposter-project/imposter-go/external/plugins/swaggerui"
+	"os"
+	"strings"
+)
+
+type SwaggerUI struct {
+	logger hclog.Logger
+}
+
+func init() {
+	specPrefixPath = os.Getenv("IMPOSTER_OPENAPI_SPEC_PATH_PREFIX")
+	if specPrefixPath == "" {
+		specPrefixPath = "/_spec"
+	}
+}
+
+func main() {
+	logger := hclog.New(&hclog.LoggerOptions{
+		Level:      hclog.Trace,
+		Output:     os.Stderr,
+		JSONFormat: true,
+	})
+
+	impl := &SwaggerUI{
+		logger: logger,
+	}
+	pluginMap := map[string]goplugin.Plugin{
+		"swaggerui": &swaggerui.SwaggerUIPlugin{Impl: impl},
+	}
+
+	logger.Debug("swaggerui plugin initialising", "path", specPrefixPath)
+
+	if logger.IsTrace() {
+		entries, err := www.ReadDir("www")
+		if err != nil {
+			panic(fmt.Errorf("failed to read static files: %v", err))
+		}
+		for _, entry := range entries {
+			logger.Trace(entry.Name())
+		}
+	}
+
+	// handshakeConfigs are used to just do a basic handshake between
+	// a plugin and host. If the handshake fails, a user-friendly error is shown.
+	// This prevents users from executing bad plugins or executing a plugin
+	// directory. It is a UX feature, not a security feature.
+	handshakeConfig := goplugin.HandshakeConfig{
+		ProtocolVersion:  1,
+		MagicCookieKey:   "HANDLER_PLUGIN",
+		MagicCookieValue: "imposter",
+	}
+
+	goplugin.Serve(&goplugin.ServeConfig{
+		HandshakeConfig: handshakeConfig,
+		Plugins:         pluginMap,
+	})
+}
+
+func (s *SwaggerUI) Configure(configs []handler.LightweightConfig) error {
+	s.logger.Trace("generating spec config")
+	if err := generateSpecConfig(configs); err != nil {
+		return fmt.Errorf("could not generate swagger UI plugin config: %w", err)
+	}
+	return nil
+}
+
+func (s *SwaggerUI) Handle(args handler.HandlerRequest) handler.HandlerResponse {
+	path := args.Path
+	if !strings.EqualFold(args.Method, "get") {
+		return handler.HandlerResponse{StatusCode: 405, Body: []byte("Method Not Allowed")}
+	}
+	if !strings.HasPrefix(path, specPrefixPath) {
+		return handler.HandlerResponse{StatusCode: 404, Body: []byte("File Not Found")}
+	}
+
+	if response := serveRawSpec(path); response != nil {
+		return *response
+	} else {
+		return serveStaticContent(path)
+	}
+}
